+++ DICUI.csproj
@@ -91,8 +91,9 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="Constants.cs" />
-    <Compile Include="DumpInformation.cs" />
-    <Compile Include="Utilities.cs" />
+    <Compile Include="Utilities\DumpInformation.cs" />
+    <Compile Include="Utilities\Validation.cs" />
+    <Compile Include="Utilities\Converters.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
