+++ Utilities/DumpInformation.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace DICUI
+namespace DICUI.Utilities
 {
     public static class DumpInformation
     {
